Allow variable sample size - with caveats
If any windows differ in their sample size, this is now allowed and
correctly enters the FEP/BAR/SOS calculations. However, inaccuracy
calculations are not currently compatible with this, so they are
disabled. There is also good reason to believe some of the display
routines don't work with this, but they at least don't crash right
now.

NB: A modest discrepancy seems to have crept into the entropy/enthalpy
calculation. The culprit is probably rounding differences due to
order of operations, as the precision difference is ~1e-6 or less.
This much precision probably should not have been reported in the
first place (I would only ever trust ~1e-3).