Fix crash when alchEquilSteps = 0
This might have been introduced recently either due to a change in the
parser or a change in the NAMD output. If no equilibration occurs, then
the normal regular expressions to identify sample fail and no data is
read.

The new code reads all valid data fields and then stores them as a
temporary buffer. If the regular expression that signifies the _end_ of
equilibration is found, then the buffer is emptied, the sample count is
reset, and equilibration steps are no longer counted. At the start of a
new window, the buffer is concatenated to the full sample list.